# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StageLink API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List events
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/buy-ticket:
    post:
      operationId: buyTicket
      tags: [events]
      summary: Buy a ticket for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketPurchaseInput"
      responses:
        "201":
          description: Ticket purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  # ── Bookings ───────────────────────────────────────────────────────────────
...
  /competitions/{competitionId}:
    get:
      operationId: getCompetition
      tags: [competitions]
      summary: Get competition details
      parameters:
        - name: competitionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /competitions/{competitionId}/enter:
    post:
      operationId: enterCompetition
      tags: [competitions]
      summary: Enter a competition
      parameters:
        - name: competitionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entered competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionEntry"
  /competitions/{competitionId}/vote/{entryId}:
    post:
      operationId: voteForEntry
      tags: [competitions]
      summary: Vote for a competition entry
      parameters:
        - name: competitionId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"

  # ── Notifications ──────────────────────────────────────────────────────────
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ArtistCard:
      type: object
      required: [id, displayName, artistType, genres, country, coverImageUrl, rating, followersCount, verified, xp, level]
      properties:
        id:
          type: integer
        displayName:
          type: string
        artistType:
          type: string
        genres:
          type: array
          items:
            type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        coverImageUrl:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        followersCount:
          type: integer
        verified:
          type: boolean
        xp:
          type: integer
        level:
          type: integer
        bookingPrice:
          type: ["integer", "null"]
...
    Event:
      type: object
      required: [id, organizerId, title, description, venue, city, country, eventDate, ticketPrice, totalTickets, soldTickets, coverImageUrl, status, featured, createdAt]
      properties:
        id:
          type: integer
        organizerId:
          type: integer
        organizerName:
          type: string
        title:
          type: string
        description:
          type: string
        venue:
          type: string
        city:
          type: string
        country:
          type: string
        eventDate:
          type: string
        ticketPrice:
          type: integer
        totalTickets:
          type: integer
        soldTickets:
          type: integer
        coverImageUrl:
          type: string
        status:
          type: string
          enum: [upcoming, live, ended, cancelled]
        featured:
          type: boolean
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ArtistCard"
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, venue, city, country, eventDate, ticketPrice, totalTickets, coverImageUrl]
      properties:
        title:
          type: string
        description:
          type: string
        venue:
          type: string
        city:
          type: string
        country:
          type: string
        eventDate:
          type: string
        ticketPrice:
          type: integer
        totalTickets:
          type: integer
        coverImageUrl:
          type: string
    Ticket:
      type: object
      required: [id, eventId, userId, qrCode, purchasedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        qrCode:
          type: string
        purchasedAt:
          type: string
    TicketPurchaseInput:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 10

    # ── Bookings ──────────────────────────────────────────────────────────
...
    Competition:
      type: object
      required: [id, title, category, description, status, prizePool, endsAt, entriesCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, voting, ended]
        prizePool:
          type: integer
        endsAt:
          type: string
        entriesCount:
          type: integer
        myEntry:
          $ref: "#/components/schemas/CompetitionEntry"
        topEntries:
          type: array
          items:
            $ref: "#/components/schemas/CompetitionEntry"
        createdAt:
          type: string
    CompetitionEntry:
      type: object
      required: [id, competitionId, artistId, artistName, votes]
      properties:
        id:
          type: integer
        competitionId:
          type: integer
        artistId:
          type: integer
        artistName:
          type: string
        artistAvatarUrl:
          type: ["string", "null"]
        submissionUrl:
          type: ["string", "null"]
        votes:
          type: integer
        rank:
          type: ["integer", "null"]
    VoteResult:
      type: object
      required: [voted, votes]
      properties:
        voted:
          type: boolean
        votes:
          type: integer

    # ── Notifications ─────────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, userId, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [follow, booking_request, booking_accepted, booking_declined, new_review, event_reminder, competition_update, ticket_purchased, message, reward]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        referenceId:
          type: ["integer", "null"]
        referenceType:
          type: ["string", "null"]
        createdAt:
          type: string
...